create project table
